requirements.txt
gradio==4.21.0
open_clip_torch @ git+https://github.com/mlfoundations/open_clip.git